some cleanup of proc allocation
