+++ openmdao/proc_allocators/proc_allocator.py
@@ -6,39 +6,6 @@ from six.moves import range
 from openmdao.utils.options_dictionary import OptionsDictionary
 
 
-class ProcAllocationError(Exception):
-    """
-    Exception raised when processor allocation fails.
-
-    Attributes
-    ----------
-    sub_idx : int
-        Index into the parent's _subsystems_allprocs list.
-    requested : int
-        Number of processes requested by the indexed subsystem.
-    remaining : int
-        Number of processes available to the indexed subsystem.
-    """
-
-    def __init__(self, sub_idx, requested, remaining):
-        """
-        Initialize all attributes.
-
-        Parameters
-        ----------
-        sub_idx : int
-            Index into the parent's _subsystems_allprocs list.
-        requested : int
-            Number of processes requested by the indexed subsystem.
-        remaining : int
-            Number of processes available to the indexed subsystem.
-        """
-        super(ProcAllocationError, self).__init__('')
-        self.sub_idx = sub_idx
-        self.requested = requested
-        self.remaining = remaining
-
-
 class ProcAllocator(object):
     """
     Algorithm for allocating processors to a given system's subsystems.
